Add functionality to support multiple username_attributes in LDAP.
When creating the user filter multiple username_attribute values will be created and OR'd together to get a filter which will return a user provided one of the username_attribute fields matches.

As an example if you wanted ActiveDirectory to find users based on either 'userPrincipalName' or 'sAMAccountName' then the following LDAP query filter will be created:

    (&(|(userPrincipalName=user.name)(sAMAccountName=user.name))(objectClass=person))

provided the following configuration

    ldap:
      filter: (objectClass=person)
      username_attribute:
        - userPrincipalName
        - sAMAccountName

particularly for Active Directory this allows users to login via either their username or their email account name.