Remove dead code
Previous to this commit, a local variable was instantiated without
side effects by the looks of it, but never used. As a wild guess, this
code was orphaned by a refactor that moved subsequent code to a
separate method that recreates the same local variable. If this isn't
dead code, then it should be refactored to make more explicit what is
going on.